TherapistPanel: Change dropdown into selection list
With the change comes a big list of smaller UI
adjustments. Unity does not have a selection list
UI on its own, so it is implemented with inspiration
from Dropdown. There remains a number of issues, but
this mainly has to do with the fact that the canvas
has been set up in the wrong resolution.