Therapist UI: Use Base Canvas of 1280x800 (16:9)
Whack-a-Mole was designed for widescreen but the
canvas in the therapist UI was assuming 800x600.
Previously this was worked around with Layout Elements
and Layout groups, but it's makes it complicated to create
UI. This commit goes over all UI elements to change their
size and layout to cater for 16:9 screens.