Feature: Add Progression Paradigm
Instead of the player waiting between each mole, if you hit a mole the next one will appear immediately.
This allows better data analysis and help to identify the player strength and weakness easier.
You can create a pattern with the HIT parameter inside of the WAIT tag, like this : WAIT:(HIT).
The game will now be based on how fast the player can be.