ModifiersManager: Offset Controllers Correctly
Before calibration was implemented, controllers
were offset by setting their world position.
However, with calibration, this results in
undefined behavior. A temporary fix was to
use localPosition, but the localPosition is
rarely aligned to world axes after calibration.
This commit implements a helper function which
reads the parents' rotation and compensates for
it, when sertting the controller's local position.