Updating Instruction Panel
We needed to update the instruction panel to see to the patient what is
the correct position to have in order to play the game. So the instruction
panel is now updating but we also need to make a fade out to the
calibration panel. However, we will wait the fade In/Out script manager
before doing this.

We added the border, we changed the picture position and the name of the UI elements.

Change of the thickness and the border color.

Change size and position of the UI elements.

Update Panel and Add fadeOut effect.