Add MoleType Column in Event Log CSV
For adding a new column in our csv log file, we need to instantiate it in the dictionaries of headers. We need to add it in the null value dictionary and in the value dictionary in order to put in the csv file the value if it's exists or just "null".